test[relay-hub]: refactor with pr suggestions
